Move commons to util/metrics
